make kubectl factory rings
