+++ pkg/kubectl/cmd/util/BUILD
@@ -14,6 +14,9 @@ go_library(
         "cached_discovery.go",
         "clientcache.go",
         "factory.go",
+        "factory_builder.go",
+        "factory_client_access.go",
+        "factory_object_mapping.go",
         "helpers.go",
         "printing.go",
         "shortcut_restmapper.go",
